Reduce Dependency on Radzen
Remaining Items to Remove:
- Popup for ItemOption editing
- Graphs (likely will keep this)

The ultimate goal of this set of changes is to reduce the number of registered services required to consume this package.

Also by switching styles over to bootstrap, this removes the requirement to have built a Radzen application to use and style this set of components.